feat: implement Sprint 5 — Transactions and MVCC
Introduces the complete ACID transaction subsystem in src/txn/:

- txid.rs: Wait-free monotonic TxID allocator (AtomicU64, SeqCst).
  Never issues TxId=0. Persists counter to system-page bytes;
  recovery bumps by 1_000_000 to leap past any pre-crash in-flight txns.

- mvcc.rs: 16-byte TupleHeader (#[repr(C)]): u32 xmin/xmax (PostgreSQL
  truncation model), u16 cid, u16 infomask, u32 next_version_ptr.
  Infomask flags: XMIN/XMAX_COMMITTED/ABORTED, HAS_NEXT_VERSION, IS_UPDATED.

- snapshot.rs: Point-in-time snapshot with PostgreSQL-style MVCC
  visibility. Lock-free hot path: binary search on sorted active list,
  cached infomask hints, self-modification transparency (owner_txid).

- state.rs: GlobalTxState — xmin/xmax as AtomicU64, active BTreeSet
  behind Mutex. begin_tx() issues TxId and builds consistent Snapshot
  atomically. Monotonic xmin advance via CAS loop on commit/abort.

- lock_table.rs: 1024-shard LockTable (Fibonacci hash). IS/IX/S/X
  modes with standard 2PL compatibility matrix. FIFO wait queues via
  crossbeam_channel bounded(1). release_all() for commit/abort cleanup.

- wound_wait.rs: Wound-wait oracle. Older requester (lower TxId) wounds
  younger holder; younger waits. Saturating u8 wound counter; immunity
  after 3 wounds prevents starvation. cleanup() on commit/abort.

- manager.rs: TransactionManager wires all subsystems. commit() flushes
  WAL before returning (durability guarantee). rollback() is best-effort.
  acquire_lock() integrates LockTable + WoundWait.

Also fixes all pre-existing clippy warnings across Sprint 1-4 modules.
240 tests pass; 0 failures.

+++ src/storage/manager.rs
@@ -83,14 +83,14 @@ impl PageManager {
 
     /// Read a page from disk into `buf`.
     pub fn read_page(&self, fs: &dyn FileSystem, page_id: PageId, buf: &mut AlignedBuffer) -> io::Result<()> {
-        let offset = page_id as u64 * PAGE_SIZE as u64;
+        let offset = page_id * PAGE_SIZE as u64;
         let handle = fs.open(&self.data_path, false)?;
         handle.read_at(buf, offset)
     }
 
     /// Write `buf` to disk at `page_id`.
     pub fn write_page(&self, fs: &dyn FileSystem, page_id: PageId, buf: &AlignedBuffer) -> io::Result<()> {
-        let offset = page_id as u64 * PAGE_SIZE as u64;
+        let offset = page_id * PAGE_SIZE as u64;
         let handle = fs.open(&self.data_path, false)?;
         handle.write_at(buf, offset)?;
         handle.sync_data()
